+++ instrumentation/github.com/gin-gonic/gin/otelgin/gintrace.go
@@ -26,13 +26,14 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/propagation"
-	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.20.0"
 	oteltrace "go.opentelemetry.io/otel/trace"
 )
 
 const (
-	tracerKey  = "otel-go-contrib-tracer"
-	tracerName = "go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin"
+	tracerKey = "otel-go-contrib-tracer"
+	// ScopeName is the instrumentation scope name.
+	ScopeName = "go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin"
 )
 
 // Middleware returns middleware that will trace incoming requests.
@@ -47,7 +48,7 @@ func Middleware(service string, opts ...Option) gin.HandlerFunc {
 		cfg.TracerProvider = otel.GetTracerProvider()
 	}
 	tracer := cfg.TracerProvider.Tracer(
-		tracerName,
+		ScopeName,
 		oteltrace.WithInstrumentationVersion(Version()),
 	)
 	if cfg.Propagators == nil {
@@ -116,7 +117,7 @@ func HTML(c *gin.Context, code int, name string, obj interface{}) {
 	}
 	if !ok {
 		tracer = otel.GetTracerProvider().Tracer(
-			tracerName,
+			ScopeName,
 			oteltrace.WithInstrumentationVersion(Version()),
 		)
 	}